requirements.txt
# Aviation RAG Dependencies
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
openai>=1.30.0
azure-search-documents>=11.4.0
azure-identity>=1.25.0
azure-core>=1.29.0

# Web framework
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.2.0

# Agent Framework runtime + observability
agent-framework>=1.0.0b260212
opentelemetry-sdk>=1.26.0
azure-monitor-opentelemetry-exporter>=1.0.0b35
